Cache invalidation mechanism
bsread_dal supports cache, because the same result can
be reused multiple times - as a positioner, as a readable,
and as a condition. We cannot just receive another message -
it a stream, so we need to cache it. If a retry happens
(the condition is not met) we need to load another message.